feat(strategy): risk-aware season_value via variance MC (Phase 2 Task 8)
_season_marginals now calls marginal_season_values_var with per-player is_rookie + fitted
params (loaded once, cached). Empty-base-roster guard (int cols) in marginal_season_values_var.
Reconciled the white-box timing test to mirror the var marginal (intent unchanged:
score = marginal - opp_cost). All season_value/strategy/tournament tests green (142).

+++ src/projections/draft/assistant/season_value.py
@@ -346,7 +346,7 @@ def marginal_season_values_var(
     avail_u = rng.random((n_sims, n_weeks, len(u_gsis)))
 
     def value_for(ids: list[str]) -> float:
-        cols = np.array([col_of[g] for g in ids])
+        cols = np.array([col_of[g] for g in ids], dtype=int)  # int even when empty (first pick)
         return _season_value_sampled(
             points_u[:, :, cols],
             avail_u[:, :, cols],
